Put shared delete child method on base class
