+++ packages/apollo-common/src/FeatureChange.ts
@@ -147,4 +147,38 @@ export abstract class FeatureChange extends AssemblySpecificChange {
       [...parentFeature.children.entries()].sort((a, b) => a[1].min - b[1].min),
     )
   }
+
+  /**
+   * Delete feature's subfeatures that match an ID and return the IDs of any
+   * sub-subfeatures that were deleted
+   * @param feature -
+   * @param featureIdToDelete -
+   * @returns - list of deleted feature IDs
+   */
+  findAndDeleteChildFeature(
+    feature: Feature,
+    featureIdToDelete: string,
+  ): string[] {
+    if (!feature.children) {
+      throw new Error(`Feature ${feature._id.toHexString()} has no children`)
+    }
+    const { _id, children } = feature
+    const child = children.get(featureIdToDelete)
+    if (child) {
+      const deletedIds = this.getChildFeatureIds(child)
+      children.delete(featureIdToDelete)
+      return deletedIds
+    }
+    for (const [, childFeature] of children) {
+      try {
+        return this.findAndDeleteChildFeature(childFeature, featureIdToDelete)
+      } catch {
+        // pass
+      }
+    }
+
+    throw new Error(
+      `Feature "${featureIdToDelete}" not found in ${_id.toHexString()}`,
+    )
+  }
 }
