+++ packages/apollo-shared/src/Changes/MergeExonsChange.ts
@@ -10,7 +10,6 @@ import {
 } from '@apollo-annotation/common'
 import { type AnnotationFeatureSnapshot } from '@apollo-annotation/mst'
 
-import { findAndDeleteChildFeature } from './DeleteFeatureChange'
 import { UndoMergeExonsChange } from './UndoMergeExonsChange'
 
 interface SerializedMergeExonsChangeBase extends SerializedFeatureChange {
@@ -84,10 +83,9 @@ export class MergeExonsChange extends FeatureChange {
       mergedExon.min = Math.min(firstExon.min, secondExon.min)
       mergedExon.max = Math.max(firstExon.max, secondExon.max)
       mergedExon.attributes = this.mergeAttributes(firstExon, secondExon)
-      const deletedIds = findAndDeleteChildFeature(
+      const deletedIds = this.findAndDeleteChildFeature(
         topLevelFeature,
         secondExon._id,
-        this,
       )
       deletedIds.push(secondExon._id)
       topLevelFeature.allIds = topLevelFeature.allIds.filter(
